fix: changed Index and Element to normal ints to fix sorting bug above 294 items

+++ includes/includes.hpp
@@ -8,8 +8,8 @@
 #include <algorithm>
 #include <string>
 
-using Element = uint16_t;
-using Index = uint16_t;
+using Element = uint32_t;
+using Index = uint32_t;
 
 constexpr uint16_t DEFAULT_WINDOW_WIDTH{ 800 };
 constexpr uint16_t DEFAULT_WINDOW_HEIGHT{ 600 };
